[update backend] add async await

+++ Back-end/Business/BPT-Service.Application/RoleService/Command/DeleteRoleAsync/DeleteRoleAsyncCommand.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Threading.Tasks;
+using BPT_Service.Application.RoleService.ViewModel;
 using BPT_Service.Model.Entities;
 using Microsoft.AspNetCore.Identity;
 
@@ -12,10 +13,32 @@ namespace BPT_Service.Application.RoleService.Command.DeleteRoleAsync
         {
             _roleManager = roleManager;
         }
-        public async Task ExecuteAsync(Guid id)
+        public async Task<CommandResult<AppRoleViewModel>> ExecuteAsync(Guid id)
         {
-            var role = await _roleManager.FindByIdAsync(id.ToString());
-            await _roleManager.DeleteAsync(role);
+            try
+            {
+                var role = await _roleManager.FindByIdAsync(id.ToString());
+                if (role != null)
+                {
+                    await _roleManager.DeleteAsync(role);
+                    return new CommandResult<AppRoleViewModel>
+                    {
+                        isValid = true
+                    };
+                }
+                return new CommandResult<AppRoleViewModel>
+                {
+                    isValid = false
+                };
+            }
+            catch (System.Exception ex)
+            {
+                return new CommandResult<AppRoleViewModel>
+                {
+                    isValid = false,
+                    errorMessage = ex.InnerException.ToString()
+                };
+            }
         }
     }
 }
